fix: fix testingz/internal.Error func
